requirements.txt
Flask==3.0.0
qdrant-client==1.7.0
openai==1.12.0
PyPDF2==3.0.1
python-dotenv==1.0.0
werkzeug==3.0.1
llama-parse==0.4.0
llama-index-core==0.10.0
twilio
